pyserver/requirements.txt
fastapi
uvicorn
diffusers[torch]
transformers
accelerate
controlnet_aux
torch
torchvision
torchaudio
opencv-python
Pillow
